refactor
- split formatting helper into their own file
- added `. presentationDetents()` helpers for scaling sheeting
- added some line limits to avoid wrapping
- report bug form padding
- scroll history event sheets
