+++ Ruddarr/Views/Movies/MovieMetadataView.swift
@@ -96,9 +96,7 @@ struct MovieMetadataView: View {
         }
         .sheet(item: $eventSheet) { event in
             MediaEventSheet(event: event)
-                .presentationDetents(
-                    event.eventType == .grabbed ? [.medium] : [.fraction(0.25)]
-                )
+                .presentationDetents(for: event)
         }
     }
