+++ Ruddarr/Views/Movies/Search/MoviePreviewView.swift
@@ -39,7 +39,7 @@ struct MoviePreviewView: View {
                         .padding(.top, -25)
                     #endif
             }
-            .presentationDetents([deviceType == .phone ? .medium : .large])
+            .presentationDetents(dynamic: [deviceType == .phone ? .medium : .large])
         }
         .alert(
             isPresented: instance.movies.errorBinding,
