+++ Ruddarr/Views/Series/Releases/SeriesReleaseSheet.swift
@@ -88,6 +88,7 @@ struct SeriesReleaseSheet: View {
             }
             .font(.subheadline)
             .foregroundStyle(.secondary)
+            .lineLimit(1)
 
             CustomFormats(tags())
         }
